Refeactored test cases Unit/Integration
