+++ src/org/bds/test/TestSuiteUnit.java
@@ -1,10 +1,8 @@
 package org.bds.test;
 
-import org.bds.test.unit.TestCasesCheckpoint;
 import org.bds.test.unit.TestCasesCommandLineOptions;
 import org.bds.test.unit.TestCasesExecutioners;
 import org.bds.test.unit.TestCasesFunctionDeclaration;
-import org.bds.test.unit.TestCasesGraph;
 import org.bds.test.unit.TestCasesInterpolate;
 import org.bds.test.unit.TestCasesLang;
 import org.bds.test.unit.TestCasesRemote;
@@ -13,8 +11,6 @@ import org.bds.test.unit.TestCasesRun;
 import org.bds.test.unit.TestCasesRun2;
 import org.bds.test.unit.TestCasesRun3;
 import org.bds.test.unit.TestCasesTail;
-import org.bds.test.unit.TestCasesTaskDetached;
-import org.bds.test.unit.TestCasesTaskImproper;
 import org.bds.test.unit.TestCasesTesting;
 import org.bds.test.unit.TestCasesVm;
 import org.junit.runner.RunWith;
@@ -37,14 +33,10 @@ import org.junit.runners.Suite.SuiteClasses;
 		TestCasesRun.class, // Running bds code
 		TestCasesRun2.class, // Running bds code
 		TestCasesRun3.class, // Running bds code (classes / object)
-		TestCasesGraph.class, // Running bds code: Task graphs and dependencies
-		TestCasesCheckpoint.class, // Running bds code: Checkpoint and recovery
 		TestCasesTesting.class, // Check bds unit testing system
 		TestCasesCommandLineOptions.class, // Check command line options
 		TestCasesRemote.class, // Accessing remote data (cloud storage)
 		TestCasesReport.class, // Report generation
-		TestCasesTaskImproper.class, // Improper tasks
-		TestCasesTaskDetached.class, // Detached tasks
 })
 public class TestSuiteUnit {
