+++ src/org/bds/test/unit/TestCasesRun2.java
@@ -1,6 +1,5 @@
 package org.bds.test.unit;
 
-import java.io.File;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
@@ -67,18 +66,6 @@ public class TestCasesRun2 extends TestCasesBase {
 		expectedValues.put("rs", "[1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0, 6.5, 7.0, 7.5, 8.0, 8.5, 9.0, 9.5, 10.0]");
 
 		runAndCheck("test/run_103.bds", expectedValues);
-	}
-
-	@Test
-	public void test104() {
-		Gpr.debug("Test");
-		runAndCheck("test/run_104.bds", "isRun", "true");
-	}
-
-	@Test
-	public void test105() {
-		Gpr.debug("Test");
-		runAndCheck("test/run_105.bds", "isRun", "false");
 	}
 
 	@Test
@@ -154,26 +141,6 @@ public class TestCasesRun2 extends TestCasesBase {
 	}
 
 	@Test
-	public void test114_parallel_function_task_calls() {
-		Gpr.debug("Test");
-		String stdout = runAndReturnStdout("test/run_114.bds");
-
-		Set<String> linesPar = new HashSet<>();
-		for (String line : stdout.split("\n"))
-			if (line.startsWith("TASK")) linesPar.add(line);
-
-		// Check
-		Assert.assertTrue("There should be 5 tasks", linesPar.size() == 5);
-	}
-
-	@Test
-	public void test115_task_dependency_using_taskId() {
-		Gpr.debug("Test");
-		String stdout = runAndReturnStdout("test/run_115.bds");
-		Assert.assertEquals("Hi 1\nBye 1\nHi 2\nBye 2\n", stdout);
-	}
-
-	@Test
 	public void test116_lineWrap_backslashId() {
 		Gpr.debug("Test");
 		String stdout = runAndReturnStdout("test/run_116.bds");
@@ -181,59 +148,9 @@ public class TestCasesRun2 extends TestCasesBase {
 	}
 
 	@Test
-	public void test117_serial_parallel_tasks() {
-		Gpr.debug("Test");
-		String expectedStdout = "Iter 1, Task 1: End\n" //
-				+ "Iter 1, Task 2: Start\n" //
-				+ "Iter 1, Task 2: End\n" //
-				+ "Iter 5, Task 1: End\n" //
-				+ "Iter 5, Task 2: Start\n" //
-				+ "Iter 5, Task 2: End\n" //
-		;;
-
-		String stdout = runAndReturnStdout("test/run_117.bds");
-
-		if (stdout.indexOf(expectedStdout) < 0) {
-			String msg = "Cannot find desired output:\n" //
-					+ "---------- Expected output ----------\n" //
-					+ expectedStdout //
-					+ "-------------- STDOUT --------------\n" //
-					+ stdout //
-			;
-			System.err.println(msg);
-			throw new RuntimeException(msg);
-		}
-	}
-
-	@Test
 	public void test118_dependency_using_path() {
 		Gpr.debug("Test");
 		runAndCheckExit("test/run_118.bds", 0);
-	}
-
-	@Test
-	public void test119_task_dependency() {
-		Gpr.debug("Test");
-
-		// Delete input file
-		String inFile = "tmp_in.txt";
-		(new File(inFile)).delete();
-
-		String expectedStdout1 = "Creating tmp_in.txt\n" //
-				+ "Running task\n" //
-				+ "Creating tmp_out.txt\n" //
-				+ "Done\n" //"
-		;
-
-		String expectedStdout2 = "Running task\n" //
-				+ "Done\n" //"
-		;
-
-		if (verbose) System.out.println("First run:");
-		runAndCheckStdout("test/run_119.bds", expectedStdout1);
-
-		if (verbose) System.out.println("\n\nSecond run:");
-		runAndCheckStdout("test/run_119.bds", expectedStdout2);
 	}
 
 	@Test
@@ -281,20 +198,6 @@ public class TestCasesRun2 extends TestCasesBase {
 		;
 
 		runAndCheckStdout("test/run_123_literals_sys.bds", output);
-	}
-
-	@Test
-	public void test123_literals_task() {
-		Gpr.debug("Test");
-
-		String output = "" //
-				// Note: This result may change if we use a different taskShell in bds.config
-				+ "task               |\t|\n" //
-				+ "task               |\t|    variable:Hello\n" //
-				+ "task               |\\t|   variable:Hello\n" //
-		;
-
-		runAndCheckStdout("test/run_123_literals_task.bds", output);
 	}
 
 	@Test
